requirements.txt
PyQt6
paramiko
GitPython
cryptography
